Install fallback EFI loader if none exist
If the user decides to install GRUB in "internal" mode, then only the 
EFI\distroname\grubx64.efi entry is created. This can cause problems if 
the system's NVRAM ever gets reset, and there is no longer a bootorder 
entry pointing to it. Therefore, we should automatically install a 
fallback loader if one doesn't already exist, to ensure the system can 
still be booted.