Set target to 0755 after rootfs extraction
If the rootfs image lacks metadata about the root directory itself, then 
the newly-installed system will have its '/' directory world-writable 
(777) by default.

This is not a fix for an osinstallgui security hole - rather, it is a 
workaround for a security hole which may exist on a distro which uses 
osinstallgui.