Add fully portable support
This aims to replace the older option of installing GRUB in internal or 
removable mode, by providing an option to fully install the OS in 
removable mode to begin with. This will set up a hybrid GRUB bootloader 
installation such that the installation can boot in BOTH Legacy BIOS and 
UEFI modes. FAQ document has also been updated to reflect this.