Only show fullport prompt on removable disks
As the prompt may be confusing, it is not in our best interest to have 
it display for internal drives by default, where it's unlikely to be 
useful anyway. Instead, we should detect if the disk is internal or 
removable (e.g., USB attached), and only prompt if it is removable.

But we should allow it to be overridden by maintainers or end users, if 
they want to have the option on internal drives anyway, and they are 
expert enough to not be confused by the prompt when it displays for an 
internal drive.