Make split-boot support more consistent
Previously split-boot support was only supported with the erase disk 
option. Make it supported on both erase disk and use existing options.

Additionally, show extra details about "what is going to be formatted" 
before the user confirms the operation in the use existing partition 
option.

Furthermore, implement a helper method that checks if a partition is 
mounted and tries to unmount it automatically if so. This reduces code 
duplication for the use-existing-partition option, as it needs to 
perform the check on all selected partitions, and the check is 
convoluted (check if mounted - try umount - check if still mounted).

Finally, add a simple check stage at the beginning of the installation 
that ensures at least one disk exists, BEFORE the user proceeds with the 
setup. This avoids the possibility that they could get through all of 
the setup and configuration, only to be told that there are no physical 
disks after the fact, and therefore they have wasted their time.