Fix no escape on locale selection
Go back two stages if the user backs out of locale selection, instead
of one. Going back one stage was fine before the initinstall stage was
added. But the addition of initinstall causes the user to get trapped
in the locale stage if they try to back out, so we need to change it to
make the locale stage go back by two if the user backs out.