Add ARGON2 support for LUKS encryption
As of the most recent bleeding-edge version of GRUB (2.14-rc1), ARGON2
is natively supported. So it is now a good time to add optional support
for it in osinstallgui. ARGON2 is considered much more resilient to GPU
based brute force attacks compared to PBKDF2. However, it comes at the
cost of less overall compatibility (GRUB, etc.), and it can also be
more resource-intensive to unlock on lower end systems.